+++ tests/spec/libsass_closed_issues/issue_338.rs
@@ -6,7 +6,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // unexepected error
 fn test() {
     assert_eq!(
         runner().ok(